Rename aas source directory to argus
