+++ src/argus/notificationprofile/models.py
@@ -8,8 +8,8 @@ from django.dispatch import receiver
 from django.utils import timezone
 from multiselectfield import MultiSelectField
 
-from aas.alert.models import Alert
-from aas.auth.models import User
+from argus.alert.models import Alert
+from argus.auth.models import User
 from .utils import AttrGetter, NestedAttrGetter
